Deduplicate CLI and server runners into a shared run-test core
